fixed pushing console
Unfortunately this introduces significant performance issues.  When
the kernel outputs the startup message char by char, it chokes the
browser.  For this reason we don't resend (now) on gdb output. This
may cause a subtle issue for now, but it's unlikely.

Into/Next also seem sluggish, but it's unclear that this change causes
it, or how to fix.